added alt attribute to img tag to meet accessibility
Image with no alt attribute
 1.1.1 Non-text Content

The image does not have an 'alt' attribute (alt="").

It’s important all images have the attribute for alternative text regardless of whether an alternative text is added.

A screen reader knows how to handle both an empty alt attribute and one with a text. If there is no attribute some screen readers will compensate and read the path to the image instead, which will often give no value to the end user.

If you are using a CMS (Content Management System), the default setting should be that an empty alt attribute is added to all images.